Added capabilities/APIs for policies for consumers * consumer can be given specific API based access to   resource groups via 'capability' request param * Updated tests

+++ test/access.py
@@ -34,7 +34,7 @@ def init_provider():
 
         try:
                 cursor.execute("update consent.role as rr set status = 'approved' from consent.users where " + " users.id = rr.user_id and users.email = 'abc.xyz@rbccps.org'")
-                cursor.execute("delete from consent.access using consent.users where access.provider_id = users.id and email = 'abc.xyz@rbccps.org'")
+                cursor.execute("delete from consent.access using consent.users where access.provider_id = users.id and email = 'abc.xyz@rbccps.org' and access_item_type = 'catalogue'")
                 conn.commit()
 
         except psycopg2.DatabaseError as error:
@@ -48,6 +48,7 @@ def reset_role(email):
                 conn.commit()
 
         except psycopg2.DatabaseError as error:
+            print(error)
             return False
         
         return True
